# BE_E_Learning/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Learning API",
        "description": "Tài liệu API cho hệ thống học trực tuyến",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/quizzes": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Lấy danh sách tất cả quizzes",
                "operationId": "c3e1008604f7b9a4737afa074426c686",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lesson_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách quizzes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Tạo mới quiz",
                "operationId": "701d22ce1cab41f2027daeeaa1791a5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "lesson_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Quiz chương 1"
                                    },
                                    "lesson_id": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Bài kiểm tra kiến thức chương 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz đã tạo"
                    }
                }
            }
        },
        "/api/quizzes/{id}": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Chi tiết quiz",
                "operationId": "4aa648545ea834d6a2b6e94d1d0df718",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết quiz"
                    }
                }
            },
            "put": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Cập nhật quiz",
                "operationId": "ff9e3388f045110c18e1be9908605a01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Quiz chương 1 - cập nhật"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nội dung mới của bài quiz"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Quiz đã cập nhật"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Xoá quiz",
                "operationId": "d721f09e0c67d7fb2b3f22404e7d4bd0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quiz đã xoá"
                    }
                }
            }
        },
        "/api/lessons/{lessonId}/quizzes": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "Lấy quiz theo bài học",
                "operationId": "d19ae4733f34630b98989940fd93e185",
                "parameters": [
                    {
                        "name": "lessonId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách quiz theo bài học"
                    }
                }
            }
        },
        "/api/quiz/submit": {
            "post": {
                "tags": [
                    "ResultQuiz"
                ],
                "summary": "Nộp bài quiz",
                "operationId": "4f1263efee7a5e4a972a1950edc06f1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quiz_id": {
                                        "type": "integer"
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "integer"
                                                },
                                                "answer_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nộp bài thành công"
                    }
                }
            }
        },
        "/api/quiz/{quizId}/my-result": {
            "get": {
                "tags": [
                    "ResultQuiz"
                ],
                "summary": "Lấy kết quả quiz của người dùng hiện tại",
                "operationId": "1cb0c07b3c30dcdc44b9de2e0dbfbe76",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách kết quả của người dùng"
                    }
                }
            }
        },
        "/api/quiz/{quizId}/results": {
            "get": {
                "tags": [
                    "ResultQuiz"
                ],
                "summary": "Lấy tất cả kết quả quiz",
                "operationId": "1e5f5f19e1857e07665a1668c2cf594e",
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách kết quả của tất cả người dùng"
                    }
                }
            }
        },
...
    },
...
}